Week 1 - Setting up Scene and Movement
Back at it again! It was fun relearning how to code again. There's still some muscle memory when it came to movement. Tripped myself up for moment when I put moveHorizontal and moveVertical in the wrong places though.

This was great and all, but if I'm really want to get into game dev, I should revise on C# and its functions, and to work on side projects to get me back into the swing of developing.