Week 7 - Level Design
Now comes the point where we take free reign with our game. I plan to make a momentum based game, so levels are gonna be big & expansive for the player to freely move around. The game is gonna be theme around exploring space ship debris, so could mess around with physics and gravity a lot.