Week 9 - Optional Modules & the Ramp
I think I accidentally glitched the system and made an immortal excel spreadsheet. Other than that, with Brendan's help and advice, figured out how to make the ramp appear after a certain amount of pickups are left, it just all has to be in the playercontroller and not in another script like a originally thought.

I highlighted the optional modules I want to add to the game. Also, perhaps with each new world, the speed and gravity of the balls changes, which could be interesting with all the powerups, booster pads and teleporters.